# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing TV series catalog",
        "contact": {
            "email": "evgenij.bx@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/otus.account.delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Delete account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otus.account.get/{id}": {
            "get": {
                "description": "Get detailed information about a account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved account",
                        "schema": {
                            "$ref": "#/definitions/handler.AccountItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/otus.account.login": {
            "post": {
                "description": "Get token by UserName",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get token by UserName",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AccountLoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/otus.account.register": {
            "post": {
                "description": "Add a new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Register new account",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AccountRegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/otus.episode.add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new episode to the catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "Create new TV episode",
                "parameters": [
                    {
                        "description": "Episode data",
                        "name": "episode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Episode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EpisodeItemResponse"
                        }
                    }
                }
            }
        },
        "/otus.episode.delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a TV episode from catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "Delete episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otus.episode.get/{id}": {
            "get": {
                "description": "Get detailed information about a TV episode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "Get episode by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved episode",
                        "schema": {
                            "$ref": "#/definitions/handler.EpisodeItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/otus.episode.update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update existing TV episode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "Update episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "episode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Episode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EpisodeItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "catalog.Episode": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "fileId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderated": {
                    "type": "boolean"
                },
                "productionDate": {
                    "type": "string"
                },
                "quality": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "seasonId": {
                    "type": "integer"
                },
                "serialId": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handler.AccountItemResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "item": {
                            "$ref": "#/definitions/model.Account"
                        }
                    }
                }
            }
        },
...
        "handler.AccountLoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "handler.AccountRegisterRequest": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Kobzar"
                },
                "last_name": {
                    "type": "string",
                    "example": ""
                },
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "name": {
                    "type": "string",
                    "example": "Evgenij"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "handler.DeleteResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "deleted": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "handler.EpisodeItemResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "item": {
                            "$ref": "#/definitions/catalog.Episode"
                        }
                    }
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "entity not found"
                }
            }
        },
...
        "model.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "example": "Kobzar"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": ""
                },
                "login": {
                    "type": "string",
                    "example": "ekobzar"
                },
                "name": {
                    "type": "string",
                    "example": "Evgenij"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}